Fix the Multer Upload Issue

+++ Backend/src/Config/cloudinary.js
@@ -27,8 +27,11 @@ const uploadOnCloudinary = async (filepath) => {
     }
     
     catch (error) {
-        fs.unlinkSync(filepath); 
+        if( filepath && fs.existsSync(filepath) ){
+            fs.unlinkSync(filepath); 
+        }
         console.log(error); 
+        return null ; 
     }
 }
